golf

+++ src/main/java/com/hook/xcs60/dao/ArrestStaffRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.hook.xcs60.model.OpsArrestStaff;
 
-public interface ArrestStaffRepository extends CrudRepository<OpsArrestStaff, Integer> {
+public interface ArrestStaffRepository extends CrudRepository<OpsArrestStaff, BigDecimal> {
 
 	@Query("SELECT o FROM OpsArrestStaff o WHERE o.staffID = :staffID") 
 	List<OpsArrestStaff> findByStaffID(@Param("staffID") BigDecimal staffID);
